unify error, debug and warning messages
Libdvbpsi used printf and fprintf to return debug and error information to the caller.
This bypasses any logging mechanism the calling application uses and does not take
multi-threaded applications into account.
The unifying of messages (error, debug or warning) allows a calling application to
filter (message level: -1, 0, 1 or 2) and handle the messages in its own way.