src/dvbpsi.c: fix segmentation fault in dvbpsi_decoder_psi_section_add()
When p->i_number == p_section->i_number AND p_prev == NULL, then the function
will crash with a segmentation fault. In this case the first element in the linked
list is going to be replaced. The pointer p_prev does not point to a valid sections,
since p is the first element in the list. To solve this case it needs to be treated
seperate.