src/dvbpsi_private.h: use correct message function.
It seems that there was a bug introduced in renaming message
dvbpsi_message in dvbpsi_private.h. This is masked (in my environment
at least) by the configure script concluding that variadic macros are
not supported (i.e. ac_cv_cpp_variadic_macros=no) if -Werror is
included in the C flags, as the fragment in configure.ac used to
determine this generates a warning and thus failure.

Signed-off-by: Jean-Paul Saman <jpsaman@videolan.org>