Fix compilation with --enable-debug with GCC 5.1
GCC 5.1.0 features stricter warnings about comparing boolean variables
with integer values, and thus compilation with --enable-debug failed
with "comparison of constant 12 with boolean expression is always true".
The leftover pieces of old code which treated integers as booleans are
now fixed, and the test suite is now compilable and runs correctly.